move queue names into models
