+++ internal/handlers/receipts.go
@@ -244,7 +244,7 @@ func QuickScan(w http.ResponseWriter, r *http.Request) {
 
 				task := asynq.NewTask(wranglerasynq.QuickScan, payloadBytes)
 
-				_, err = wranglerasynq.EnqueueTask(task, wranglerasynq.QuickScanQueue)
+				_, err = wranglerasynq.EnqueueTask(task, models.QuickScanQueue)
 				if err != nil {
 					return http.StatusInternalServerError, err
 				}
