+++ internal/models/asynq_queue_configuration.go
@@ -1,8 +1,6 @@
 package models
 
-import "receipt-wrangler/api/internal/wranglerasynq"
-
 type AsynqQueueConfiguration struct {
-	QueueName wranglerasynq.QueueName
+	QueueName QueueName
 	Priority  int
 }
