+++ internal/wranglerasynq/task_enqueue.go
@@ -2,14 +2,15 @@ package wranglerasynq
 
 import (
 	"github.com/hibiken/asynq"
+	"receipt-wrangler/api/internal/models"
 	"receipt-wrangler/api/internal/repositories"
 )
 
-func EnqueueTask(task *asynq.Task, queue QueueName) (*asynq.TaskInfo, error) {
+func EnqueueTask(task *asynq.Task, queue models.QueueName) (*asynq.TaskInfo, error) {
 	client := repositories.GetAsynqClient()
 	return client.Enqueue(task, asynq.MaxRetry(3), asynq.Queue(string(queue)))
 }
 
-func RegisterTask(cronspec string, task *asynq.Task, queue QueueName, maxRetry int) (string, error) {
+func RegisterTask(cronspec string, task *asynq.Task, queue models.QueueName, maxRetry int) (string, error) {
 	return scheduler.Register(cronspec, task, asynq.MaxRetry(maxRetry), asynq.Queue(string(queue)))
 }
